+ Actualización de Script para PROD

+++ start_containers.sh
@@ -9,9 +9,21 @@ then
     exit 1
 fi
 
+# Ensure correct permissions for directories and files
+echo "[JACKMOON-DEV] Setting permissions for directories and files..."
+chmod -R u+w ./docker
+chmod -R u+w ./sql
+chmod u+w ./clean.sh
+
+if [ $? -ne 0 ]; then
+    echo "[JACKMOON-DEV] Error while setting permissions."
+    exit 1
+fi
+
 # Stop and remove existing containers
 echo "[JACKMOON-DEV] Stopping and removing existing containers..."
 docker-compose down
+sleep 10  # Adding a delay to ensure proper shutdown
 
 if [ $? -ne 0 ]; then
     echo "[JACKMOON-DEV] Error while stopping and removing containers."
